We've added two new endpoints for DRC to keep us updated on the files.
Whenever we receive a request from DRC, our API will forward it to the maat-api for downstream processing.

As part of this effort, we've also integrated a woremock to test the maat-api response and our API's behavior.

Based on the response, our API will return a string message to DRC.
Story: DCES-285